fix(parse): fix stripping of flag with no parameters
Flags like `#qj-id: 24` are stripped by the regex, but flags that have no parameter like `qj-show-json` and `#qj-hide-id` are not.

The reason is that the regex expect to match the character sequence `: ` in it. But these flags do not match it and are not stripped from the input data.

This PR updates the regex to make the character sequence `: ` optional, thus making the flags match.